centre of mass calculation and rotation around com support
